Fixed bugs and improved behavior

+++ app/src/main/java/ru/tech/imageresizershrinker/batch_resize/BatchResizeScreen.kt
@@ -325,7 +325,7 @@ fun BatchResizeScreen(
                                     Text(
                                         stringResource(
                                             R.string.size,
-                                            byteCount(bitmapInfo.size)
+                                            byteCount(bitmapInfo.size.toLong())
                                         )
                                     )
                                 } else {
